Initial commit of eu-central-1 customizations
Previously, you could use the classic endpoint for DNS compatible
buckets (bucket.s3.amazonaws.com), and it would automatically route
you to the correct region.

However, with signature version 4, you now need to know the region
of the bucket so you can properly sign the request.

eu-central-1 is the first signature version 4 only region, and requires
us to know the region.

So this commit adds a customization that will take any error response
from s3, and see if the error message says "you need to use sigv4".
If it does, we try to figure out the bucket location via DNS and
switch the endpoint and auth classes over to s3v4.

The reason this fix is not going in right now is because this approach
does not work for HEAD bucket/object, which is used for things like
checking if an object exists, or in the case of a CLI, getting an
object's size to determine if we want to download the object
in ranges.  Rather than have a partially complete solution, we can
just encourage customers to specify a region when using buckets in
eu-central-1.  Specifying an exact region works 100% of the time.